feat: Add TrendingPools component to dashboard page
This commit adds the TrendingPools component to the dashboard page in the frontend app. The TrendingPools component fetches data from the GraphQL API using the useQuery hook from @apollo/client and displays a list of trending pools. This enhances the functionality and user experience of the dashboard by providing valuable information to the users.