refactor: Update DeployDSC script to return DSC instance
The DeployDSC script has been updated to return the DSC instance after executing the deployment process. This allows for easier access to the deployed contract in subsequent operations.

Refactor the `run` function to include a return type of `DSC` and add a return statement at the end of the function.

Fixes #27